pyproject.toml
[project]
name = "dsubot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "disnake>=2.10.1",
    "python-dotenv>=1.0.1",
]
